initial cleanup
