+++ functions/googleHome/colorDictionary.js
