+++ package.json
@@ -2,7 +2,7 @@
   "name": "functions",
   "description": "Cloud Functions for Firebase",
   "engines": {
-    "node": "8"
+    "node": "8.12.0"
   },
   "scripts": {
     "dev": "firebase serve --only functions",
@@ -19,8 +19,7 @@
     "express": "^4.16.3",
     "firebase-admin": "~5.10.0",
     "firebase-functions": "^2.0.5",
-    "glob": "^7.1.3",
-    "jwt-decode": "^2.2.0"
+    "glob": "^7.1.3"
   },
   "devDependencies": {
     "eslint": "^4.12.0",
